✨ Add file to deploy and manage OpenBao identity groups
Add base modules files to deploy OpenBao groups.

Since a single group should be manage by the module, every attache policies,
identity entity or ideneity groups are attache with the `exclusive` parameter
set to true.

If another team deploy the same group with other attachment, this will lead to
infrastructure drift.

This beahaviour is expective as I consider if this happens, it means something
went wrong, either someone gain access to the vault or a another team is making
update to the group deployed by this module without using dedicate
implementation.